requirements.txt
# Django and REST framework
Django>=5.1.5
djangorestframework>=3.14.0
django-filter>=24.1

# JWT authentication
djangorestframework-simplejwt>=5.4.0

# CORS support
django-cors-headers>=4.6.0

# For image processing
Pillow>=10.2.0

psycopg2-binary>=2.9.9
boto3>=1.34.0
python-jose>=3.3.0
django-storages>=1.14.2
firebase-admin>=6.4.0 

drf-nested-routers==0.94.1
django-environ==0.11.2
django-storages[boto3]